# docs/swagger.yaml
definitions:
  models.Appointment:
    properties:
      completed:
        default: false
        type: boolean
      created_at:
        type: string
      date:
        type: string
      duration:
        type: integer
      id:
        type: integer
      notes:
        type: string
      pet_id:
        type: integer
      reason:
        type: string
      updated_at:
        type: string
    required:
    - date
    - duration
    - pet_id
    - reason
    type: object
...
  models.Pet:
    properties:
      appointments:
        items:
          $ref: '#/definitions/models.Appointment'
        type: array
      birth_date:
        type: string
      breed:
        type: string
      client_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      species:
        type: string
      updated_at:
        type: string
      weight:
        type: number
    required:
    - client_id
    - name
    - species
    type: object
info:
  contact: {}
paths:
...
  /api/v1/clients/{id}/pets:
    get:
      consumes:
      - application/json
      description: Obtiene todas las mascotas de un cliente específico
      parameters:
      - description: ID del cliente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Pet'
            type: array
        "400":
          description: Formato de ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Cliente no encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      summary: Obtiene mascotas por cliente
      tags:
      - Pets
  /api/v1/pets:
    get:
      consumes:
      - application/json
      description: Obtiene todas las mascotas o filtra por cliente si se especifica
        el parámetro `client_id`.
      parameters:
      - description: ID del cliente
        in: query
        name: client_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Pet'
            type: array
        "400":
          description: Formato de ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Cliente no encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      summary: Obtiene mascotas
      tags:
      - Pets
    post:
      consumes:
      - application/json
      description: Crea una nueva mascota
      parameters:
      - description: Datos de la mascota
        in: body
        name: pet
        required: true
        schema:
          $ref: '#/definitions/models.Pet'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Pet'
        "400":
          description: Error en los datos enviados
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Cliente no encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      summary: Crea mascota
      tags:
      - Pets
  /api/v1/pets/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina una mascota por su ID
      parameters:
      - description: ID de la mascota
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Mascota eliminada correctamente
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Formato de ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Mascota no encontrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      summary: Elimina mascota
      tags:
      - Pets
    get:
      consumes:
      - application/json
      description: Obtiene los detalles de una mascota específica por su ID
      parameters:
      - description: ID de la mascota
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Pet'
        "400":
          description: Formato de ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Mascota no encontrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      summary: Obtiene mascota por ID
      tags:
      - Pets
    put:
      consumes:
      - application/json
      description: Modifica una mascota total o parcialmente
      parameters:
      - description: ID de la mascota
        in: path
        name: id
        required: true
        type: integer
      - description: Datos de la mascota
        in: body
        name: pet
        required: true
        schema:
          $ref: '#/definitions/models.Pet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Pet'
        "400":
          description: Error en los datos enviados
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Mascota o cliente no encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      summary: Modifica mascota
      tags:
      - Pets
...
swagger: "2.0"